write new example for learning task
